requirements.txt
# Primary Dependencies
PyJWT==2.10.1
python-dotenv==1.1.1
requests==2.32.5

# Packaging Dependencies
# black==25.9.0
# build==1.3.0
# packaging==25.0
# pip==25.2
# setuptools==80.9.0
# twine==6.2.0
# wheel==0.45.1

# Testing Dependencies
# pytest==8.4.2
# coverage==7.10.7
# pyflakes==3.4.0
# pytest-cov==7.0.0
# pytest-flakes==4.0.5
# pytest-spec==5.1.0

# Subdependencies
# certifi==2025.8.3
# charset-normalizer==3.4.3
# click==8.3.0
# docutils==0.22.2
# dotenv==0.9.9
# id==1.5.0
# idna==3.10
# iniconfig==2.1.0
# jaraco.classes==3.4.0
# jaraco.context==6.0.1
# jaraco.functools==4.3.0
# keyring==25.6.0
# markdown-it-py==4.0.0
# mdurl==0.1.2
# more-itertools==10.8.0
# mypy_extensions==1.1.0
# nh3==0.3.0
# pathspec==0.12.1
# platformdirs==4.4.0
# pluggy==1.6.0
# Pygments==2.19.2
# pyproject_hooks==1.2.0
# pytokens==0.1.10
# readme_renderer==44.0
# requests-toolbelt==1.0.0
# rfc3986==2.0.0
# rich==14.1.0
# urllib3==2.5.0